+++ package.json
@@ -16,8 +16,10 @@
     "@hookform/resolvers": "^3.3.4",
     "@mui/icons-material": "^5.15.10",
     "@mui/material": "^5.15.10",
+    "@mui/x-date-pickers": "^6.19.4",
     "@reduxjs/toolkit": "^2.2.1",
     "axios": "^1.6.7",
+    "dayjs": "^1.11.10",
     "dotenv": "^16.4.1",
     "firebase": "^10.6.0",
     "pdf-lib": "^1.17.1",
